\name{A-2695}
\alias{A-2695}
\docType{data}
\title{
An example of Alpha Beta k dataset
}
\description{
Contains data to test the \code{read.abk} and \code{write.abk} functions.
}
\usage{data(a2695)}
\format{
  A data frame with 100 features (rows) on 22 cases (columns).
}
\source{
}
\references{
}
\examples{
data(a2695)
a2695
}
\keyword{datasets}


