\name{cibm-utils-package}
\alias{cibm-utils-package}
\alias{cibm-utils}
\docType{package}
\title{
CIBM's R utility functions
}
\description{
Contains the R implementation of a set of functions commonly used by CIBM members.
}
\details{
\tabular{ll}{
Package: \tab cibm-utils\cr
Type: \tab Package\cr
Version: \tab 0.1.5\cr
Date: \tab 2013-08-06\cr
License: \tab MIT\cr
}
}
\author{
Renato Vimieiro and Carlos Riveros

Maintainer: Renato Vimieiro <renato.vimieiro@newcastle.edu.au>, Carlos Riveros
    <carlos.riveros@newcastle.edu.au>
}
\keyword{ CIBM }
\keyword{ clustering }
\keyword{ distance measure }
