\docType{methods}
\name{labels}
\alias{features}
\alias{features,cibm.data,cibm.data-methods}
\alias{labels}
\alias{labels,cibm.data,cibm.data-methods}
\alias{labels<-}
\alias{properties}
\alias{properties<-}
\alias{samples}
\alias{samples,cibm.data,cibm.data-methods}
\title{Accessor functions}
\arguments{
  \item{object}{\code{cibm.base}. An object to retrieve set
  information.}

  \item{value}{Data to set \code{labels}}

  \item{object}{Of class \code{cibm.base}. An object to
  retrieve set information.}

  \item{object}{Of class \code{cibm.base}. An object to
  retrieve set information.}
}
\description{
  Accessor functions
}

